% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrateIt.R
\name{integrateIt}
\alias{integrateIt}
\alias{integrateIt,ANY-method}
\title{Estimates the integral of a mathematical function}
\usage{
integrateIt(rule = "Trap", x, y, a, b)
}
\arguments{
\item{rule}{Which rule, either 'Trap' or 'Simp', one would like to use to estimate the integral}

\item{x}{A vector of values}

\item{y}{A vector of evaluated values}

\item{a}{Lower bound of the integral}

\item{b}{Upper bound of the integral}
}
\value{
An object of either class `Trapezoid' or class 'Simpson' that contains
\itemize{
\item \code{x} A vector of values
\item \code{y} A vector of evaluated values
\item \code{a} Lower bound of the integral
\item \code{b} Upper bound of the integral
\item \code{estInt} Estimated integral
}
}
\description{
Estimates the integral of a mathematical funciton using either the trapezoidal rule or Simpson's rule
}
\examples{

example.x <- 1:25
example.y <- (example.x)^2
integrateIt("Trap", example.x, example.y, 4, 22)
integrateIt("Simp", example.x, example.y, 4, 22)

}
\author{
Jacob H. Hample: \email{jacob.hample@wustl.edu}
}
\seealso{
\code{\link{Trapezoid}}, \code{\link{Simpson}}
}

