% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Trapezoid.R
\docType{class}
\name{Trapezoid-class}
\alias{Trapezoid-class}
\alias{initialize,Trapezoid-method}
\alias{plot,Trapezoid-method}
\alias{print,Trapezoid-method}
\title{A Trapezoid object}
\description{
Objects of class \code{Trapezoid} are integrals estimated using the trapezoidal rule
}
\details{
An object of the class Trapezoid has the following slots:
\itemize{
\item \code{x} A vector of values
\item \code{y} A vector of evaluated values
\item \code{a} Lower bound of the integral
\item \code{b} Upper bound of the integral
\item \code{estInt} Estimated integral
}
}
\author{
Jacob H. Hample: \email{jacob.hample@wustl.edu}
}

